implemented Listmanager class, shifted the print list function into this class
